requirements.txt
python==3.13 
torch 
torchvision 
pytorch-lightning 
numpy